@@ -25,11 +25,11 @@
     "clsx": "^2.1.1",
     "date-fns": "^4.1.0",
     "lucide-react": "^0.564.0",
-    "next": "^15.1.7",
+    "next": "^16.1.6",
     "next-pwa": "^5.6.0",
     "next-themes": "^0.4.6",
-    "react": "^19.0.0",
-    "react-dom": "^19.0.0",
+    "react": "^19.2.4",
+    "react-dom": "^19.2.4",
     "react-hook-form": "^7.71.1",
     "react-intersection-observer": "^10.0.2",
     "sonner": "^2.0.7",
@@ -45,12 +45,12 @@
     "@types/react": "^19",
     "@types/react-dom": "^19",
     "eslint": "^9",
-    "eslint-config-next": "^15.1.7",
+    "eslint-config-next": "^16.1.6",
     "tailwindcss": "^4",
     "typescript": "^5"
   },
   "description": "A robust, full-stack community issue reporting application built with modern web technologies. JomReport empowers communities to report local issues (like potholes, broken streetlights) and track their resolution in real-time.",
-  "main": "eslint.config.js",
+  "main": "eslint.config.mjs",
   "repository": {
     "type": "git",
     "url": "git+https://github.com/Ajwdxr/JomReportv2.git"
@@ -58,7 +58,7 @@
   "keywords": [],
   "author": "",
   "license": "ISC",
-  "type": "commonjs",
+  "type": "module",
   "bugs": {
     "url": "https://github.com/Ajwdxr/JomReportv2/issues"
   },
