@@ -1,4 +1,9 @@
-const { FlatCompat } = require("@eslint/eslintrc");
+import { dirname } from "path";
+import { fileURLToPath } from "url";
+import { FlatCompat } from "@eslint/eslintrc";
+
+const __filename = fileURLToPath(import.meta.url);
+const __dirname = dirname(__filename);
 
 const compat = new FlatCompat({
   baseDirectory: __dirname,
@@ -28,4 +33,4 @@ const eslintConfig = [
   }
 ];
 
-module.exports = eslintConfig;
+export default eslintConfig;
