@@ -1,1 +1,1 @@
-if(!self.define){let e,s={};const i=(i,t)=>(i=new URL(i+".js",t).href,s[i]||new Promise(s=>{if("document"in self){const e=document.createElement("script");e.src=i,e.onload=s,document.head.appendChild(e)}else e=i,importScripts(i),s()}).then(()=>{let e=s[i];if(!e)throw new Error(`Module ${i} didn’t register its module`);return e}));self.define=(t,a)=>{const n=e||("document"in self?document.currentScript.src:"")||location.href;if(s[n])return;let c={};const r=e=>i(e,n),u={module:{uri:n},exports:c,require:r};s[n]=Promise.all(t.map(e=>u[e]||r(e))).then(e=>(a(...e),c))}}define(["./workbox-4754cb34"],function(e){"use strict";importScripts(),self.skipWaiting(),e.clientsClaim(),e.precacheAndRoute([{url:"/_next/app-build-manifest.json",revision:"666d86bddf47e27e045585e544a477c1"},{url:"/_next/static/Emg1NlNQ_B_i-u3rtd_Hz/_buildManifest.js",revision:"a05f9edc23546eb1b60a411d762fcdcc"},{url:"/_next/static/Emg1NlNQ_B_i-u3rtd_Hz/_ssgManifest.js",revision:"b6652df95db52feb4daf4eca35380933"},{url:"/_next/static/chunks/1-71ed8eee042ada87.js",revision:"Emg1NlNQ_B_i-u3rtd_Hz"},{url:"/_next/static/chunks/12-4536054efc5f0b17.js",revision:"Emg1NlNQ_B_i-u3rtd_Hz"},{url:"/_next/static/chunks/203.2b4c1ee4fbe3a7cf.js",revision:"2b4c1ee4fbe3a7cf"},{url:"/_next/static/chunks/218.57a830a2c55ba802.js",revision:"57a830a2c55ba802"},{url:"/_next/static/chunks/276-a1f3d6b6d1e9f44e.js",revision:"Emg1NlNQ_B_i-u3rtd_Hz"},{url:"/_next/static/chunks/35-d9bfbfe3201f31a8.js",revision:"Emg1NlNQ_B_i-u3rtd_Hz"},{url:"/_next/static/chunks/415-9d4e8f02c38f650a.js",revision:"Emg1NlNQ_B_i-u3rtd_Hz"},{url:"/_next/static/chunks/486-8a90f5a0d633d27f.js",revision:"Emg1NlNQ_B_i-u3rtd_Hz"},{url:"/_next/static/chunks/4bd1b696-439af242d2251dc4.js",revision:"Emg1NlNQ_B_i-u3rtd_Hz"},{url:"/_next/static/chunks/517-a7f406c93a55d7db.js",revision:"Emg1NlNQ_B_i-u3rtd_Hz"},{url:"/_next/static/chunks/521-8eb1004201715970.js",revision:"Emg1NlNQ_B_i-u3rtd_Hz"},{url:"/_next/static/chunks/554-45e3ac9c3a34067b.js",revision:"Emg1NlNQ_B_i-u3rtd_Hz"},{url:"/_next/static/chunks/671-51c40a9f231174b7.js",revision:"Emg1NlNQ_B_i-u3rtd_Hz"},{url:"/_next/static/chunks/710-83ed7f0ee7cd8aaf.js",revision:"Emg1NlNQ_B_i-u3rtd_Hz"},{url:"/_next/static/chunks/76-765dbf9527983ffd.js",revision:"Emg1NlNQ_B_i-u3rtd_Hz"},{url:"/_next/static/chunks/796-794b3f869bd8d0aa.js",revision:"Emg1NlNQ_B_i-u3rtd_Hz"},{url:"/_next/static/chunks/907-721824828a93b94b.js",revision:"Emg1NlNQ_B_i-u3rtd_Hz"},{url:"/_next/static/chunks/970-9196d30092cc342e.js",revision:"Emg1NlNQ_B_i-u3rtd_Hz"},{url:"/_next/static/chunks/app/_not-found/page-76356ec7aa00af5e.js",revision:"Emg1NlNQ_B_i-u3rtd_Hz"},{url:"/_next/static/chunks/app/admin/page-9debc24a359792a4.js",revision:"Emg1NlNQ_B_i-u3rtd_Hz"},{url:"/_next/static/chunks/app/api/geocode/route-6180abf3330be1a4.js",revision:"Emg1NlNQ_B_i-u3rtd_Hz"},{url:"/_next/static/chunks/app/auth/callback/route-48d2cb284cec7c54.js",revision:"Emg1NlNQ_B_i-u3rtd_Hz"},{url:"/_next/static/chunks/app/layout-2776c0b43f786c35.js",revision:"Emg1NlNQ_B_i-u3rtd_Hz"},{url:"/_next/static/chunks/app/login/page-922eff23f656cf51.js",revision:"Emg1NlNQ_B_i-u3rtd_Hz"},{url:"/_next/static/chunks/app/notifications/page-c4b0da7a402656f9.js",revision:"Emg1NlNQ_B_i-u3rtd_Hz"},{url:"/_next/static/chunks/app/page-d0c796d547f1d5d9.js",revision:"Emg1NlNQ_B_i-u3rtd_Hz"},{url:"/_next/static/chunks/app/profile/%5Bid%5D/page-92ac853a5e464f30.js",revision:"Emg1NlNQ_B_i-u3rtd_Hz"},{url:"/_next/static/chunks/app/profile/page-29f8a979cf3639d4.js",revision:"Emg1NlNQ_B_i-u3rtd_Hz"},{url:"/_next/static/chunks/app/report/%5Bid%5D/page-9d1fefe332089e85.js",revision:"Emg1NlNQ_B_i-u3rtd_Hz"},{url:"/_next/static/chunks/app/report/new/page-8a517feb676aed3f.js",revision:"Emg1NlNQ_B_i-u3rtd_Hz"},{url:"/_next/static/chunks/app/trending/page-f688d0ea98e0402e.js",revision:"Emg1NlNQ_B_i-u3rtd_Hz"},{url:"/_next/static/chunks/framework-6b27c2b7aa38af2d.js",revision:"Emg1NlNQ_B_i-u3rtd_Hz"},{url:"/_next/static/chunks/main-0ccbf0ddd675be13.js",revision:"Emg1NlNQ_B_i-u3rtd_Hz"},{url:"/_next/static/chunks/main-app-f83ad3163740aa35.js",revision:"Emg1NlNQ_B_i-u3rtd_Hz"},{url:"/_next/static/chunks/pages/_app-04c695bc05fa7935.js",revision:"Emg1NlNQ_B_i-u3rtd_Hz"},{url:"/_next/static/chunks/pages/_error-da21db4e7d4f6e09.js",revision:"Emg1NlNQ_B_i-u3rtd_Hz"},{url:"/_next/static/chunks/polyfills-42372ed130431b0a.js",revision:"846118c33b2c0e922d7b3a7676f81f6f"},{url:"/_next/static/chunks/webpack-3f1db06c145d9582.js",revision:"Emg1NlNQ_B_i-u3rtd_Hz"},{url:"/_next/static/css/5e8da8aef87db279.css",revision:"5e8da8aef87db279"},{url:"/_next/static/media/4cf2300e9c8272f7-s.p.woff2",revision:"18bae71b1e1b2bb25321090a3b563103"},{url:"/_next/static/media/747892c23ea88013-s.woff2",revision:"a0761690ccf4441ace5cec893b82d4ab"},{url:"/_next/static/media/8d697b304b401681-s.woff2",revision:"cc728f6c0adb04da0dfcb0fc436a8ae5"},{url:"/_next/static/media/93f479601ee12b01-s.p.woff2",revision:"da83d5f06d825c5ae65b7cca706cb312"},{url:"/_next/static/media/9610d9e46709d722-s.woff2",revision:"7b7c0ef93df188a852344fc272fc096b"},{url:"/_next/static/media/ba015fad6dcf6784-s.woff2",revision:"8ea4f719af3312a055caf09f34c89a77"},{url:"/file.svg",revision:"d09f95206c3fa0bb9bd9fefabfd0ea71"},{url:"/globe.svg",revision:"2aaafa6a49b6563925fe440891e32717"},{url:"/icons/icon.svg",revision:"4912854fbced0ec54f13367855034d23"},{url:"/jomreport.png",revision:"b08d712a0bb1ffb3aca40450f5463e54"},{url:"/manifest.json",revision:"6c8c6fa4340ae62fb166acc9c4b12b47"},{url:"/next.svg",revision:"8e061864f388b47f33a1c3780831193e"},{url:"/vercel.svg",revision:"c0af2f507b369b085b35ef4bbe3bcf1e"},{url:"/window.svg",revision:"a2760511c65806022ad20adf74370ff3"}],{ignoreURLParametersMatching:[]}),e.cleanupOutdatedCaches(),e.registerRoute("/",new e.NetworkFirst({cacheName:"start-url",plugins:[{cacheWillUpdate:async({request:e,response:s,event:i,state:t})=>s&&"opaqueredirect"===s.type?new Response(s.body,{status:200,statusText:"OK",headers:s.headers}):s}]}),"GET"),e.registerRoute(/^https:\/\/fonts\.(?:gstatic)\.com\/.*/i,new e.CacheFirst({cacheName:"google-fonts-webfonts",plugins:[new e.ExpirationPlugin({maxEntries:4,maxAgeSeconds:31536e3})]}),"GET"),e.registerRoute(/^https:\/\/fonts\.(?:googleapis)\.com\/.*/i,new e.StaleWhileRevalidate({cacheName:"google-fonts-stylesheets",plugins:[new e.ExpirationPlugin({maxEntries:4,maxAgeSeconds:604800})]}),"GET"),e.registerRoute(/\.(?:eot|otf|ttc|ttf|woff|woff2|font.css)$/i,new e.StaleWhileRevalidate({cacheName:"static-font-assets",plugins:[new e.ExpirationPlugin({maxEntries:4,maxAgeSeconds:604800})]}),"GET"),e.registerRoute(/\.(?:jpg|jpeg|gif|png|svg|ico|webp)$/i,new e.StaleWhileRevalidate({cacheName:"static-image-assets",plugins:[new e.ExpirationPlugin({maxEntries:64,maxAgeSeconds:86400})]}),"GET"),e.registerRoute(/\/_next\/image\?url=.+$/i,new e.StaleWhileRevalidate({cacheName:"next-image",plugins:[new e.ExpirationPlugin({maxEntries:64,maxAgeSeconds:86400})]}),"GET"),e.registerRoute(/\.(?:mp3|wav|ogg)$/i,new e.CacheFirst({cacheName:"static-audio-assets",plugins:[new e.RangeRequestsPlugin,new e.ExpirationPlugin({maxEntries:32,maxAgeSeconds:86400})]}),"GET"),e.registerRoute(/\.(?:mp4)$/i,new e.CacheFirst({cacheName:"static-video-assets",plugins:[new e.RangeRequestsPlugin,new e.ExpirationPlugin({maxEntries:32,maxAgeSeconds:86400})]}),"GET"),e.registerRoute(/\.(?:js)$/i,new e.StaleWhileRevalidate({cacheName:"static-js-assets",plugins:[new e.ExpirationPlugin({maxEntries:32,maxAgeSeconds:86400})]}),"GET"),e.registerRoute(/\.(?:css|less)$/i,new e.StaleWhileRevalidate({cacheName:"static-style-assets",plugins:[new e.ExpirationPlugin({maxEntries:32,maxAgeSeconds:86400})]}),"GET"),e.registerRoute(/\/_next\/data\/.+\/.+\.json$/i,new e.StaleWhileRevalidate({cacheName:"next-data",plugins:[new e.ExpirationPlugin({maxEntries:32,maxAgeSeconds:86400})]}),"GET"),e.registerRoute(/\.(?:json|xml|csv)$/i,new e.NetworkFirst({cacheName:"static-data-assets",plugins:[new e.ExpirationPlugin({maxEntries:32,maxAgeSeconds:86400})]}),"GET"),e.registerRoute(({url:e})=>{if(!(self.origin===e.origin))return!1;const s=e.pathname;return!s.startsWith("/api/auth/")&&!!s.startsWith("/api/")},new e.NetworkFirst({cacheName:"apis",networkTimeoutSeconds:10,plugins:[new e.ExpirationPlugin({maxEntries:16,maxAgeSeconds:86400})]}),"GET"),e.registerRoute(({url:e})=>{if(!(self.origin===e.origin))return!1;return!e.pathname.startsWith("/api/")},new e.NetworkFirst({cacheName:"others",networkTimeoutSeconds:10,plugins:[new e.ExpirationPlugin({maxEntries:32,maxAgeSeconds:86400})]}),"GET"),e.registerRoute(({url:e})=>!(self.origin===e.origin),new e.NetworkFirst({cacheName:"cross-origin",networkTimeoutSeconds:10,plugins:[new e.ExpirationPlugin({maxEntries:32,maxAgeSeconds:3600})]}),"GET")});
+if(!self.define){let e,i={};const s=(s,a)=>(s=new URL(s+".js",a).href,i[s]||new Promise(i=>{if("document"in self){const e=document.createElement("script");e.src=s,e.onload=i,document.head.appendChild(e)}else e=s,importScripts(s),i()}).then(()=>{let e=i[s];if(!e)throw new Error(`Module ${s} didn’t register its module`);return e}));self.define=(a,n)=>{const t=e||("document"in self?document.currentScript.src:"")||location.href;if(i[t])return;let c={};const r=e=>s(e,t),f={module:{uri:t},exports:c,require:r};i[t]=Promise.all(a.map(e=>f[e]||r(e))).then(e=>(n(...e),c))}}define(["./workbox-4754cb34"],function(e){"use strict";importScripts(),self.skipWaiting(),e.clientsClaim(),e.precacheAndRoute([{url:"/_next/app-build-manifest.json",revision:"22ee886ad8d8470934b40a66bfc461ac"},{url:"/_next/static/U6_6AiTyhfp9Ki-K_78RL/_buildManifest.js",revision:"a05f9edc23546eb1b60a411d762fcdcc"},{url:"/_next/static/U6_6AiTyhfp9Ki-K_78RL/_ssgManifest.js",revision:"b6652df95db52feb4daf4eca35380933"},{url:"/_next/static/chunks/1-71ed8eee042ada87.js",revision:"U6_6AiTyhfp9Ki-K_78RL"},{url:"/_next/static/chunks/12-4536054efc5f0b17.js",revision:"U6_6AiTyhfp9Ki-K_78RL"},{url:"/_next/static/chunks/203.2b4c1ee4fbe3a7cf.js",revision:"2b4c1ee4fbe3a7cf"},{url:"/_next/static/chunks/218.57a830a2c55ba802.js",revision:"57a830a2c55ba802"},{url:"/_next/static/chunks/276-46440f2abf0735b9.js",revision:"U6_6AiTyhfp9Ki-K_78RL"},{url:"/_next/static/chunks/35-d9bfbfe3201f31a8.js",revision:"U6_6AiTyhfp9Ki-K_78RL"},{url:"/_next/static/chunks/415-9d4e8f02c38f650a.js",revision:"U6_6AiTyhfp9Ki-K_78RL"},{url:"/_next/static/chunks/486-8a90f5a0d633d27f.js",revision:"U6_6AiTyhfp9Ki-K_78RL"},{url:"/_next/static/chunks/4bd1b696-439af242d2251dc4.js",revision:"U6_6AiTyhfp9Ki-K_78RL"},{url:"/_next/static/chunks/517-a7f406c93a55d7db.js",revision:"U6_6AiTyhfp9Ki-K_78RL"},{url:"/_next/static/chunks/521-8eb1004201715970.js",revision:"U6_6AiTyhfp9Ki-K_78RL"},{url:"/_next/static/chunks/554-45e3ac9c3a34067b.js",revision:"U6_6AiTyhfp9Ki-K_78RL"},{url:"/_next/static/chunks/671-51c40a9f231174b7.js",revision:"U6_6AiTyhfp9Ki-K_78RL"},{url:"/_next/static/chunks/710-83ed7f0ee7cd8aaf.js",revision:"U6_6AiTyhfp9Ki-K_78RL"},{url:"/_next/static/chunks/76-765dbf9527983ffd.js",revision:"U6_6AiTyhfp9Ki-K_78RL"},{url:"/_next/static/chunks/796-794b3f869bd8d0aa.js",revision:"U6_6AiTyhfp9Ki-K_78RL"},{url:"/_next/static/chunks/907-721824828a93b94b.js",revision:"U6_6AiTyhfp9Ki-K_78RL"},{url:"/_next/static/chunks/970-9196d30092cc342e.js",revision:"U6_6AiTyhfp9Ki-K_78RL"},{url:"/_next/static/chunks/app/_not-found/page-76356ec7aa00af5e.js",revision:"U6_6AiTyhfp9Ki-K_78RL"},{url:"/_next/static/chunks/app/admin/page-8bf9818a40ef8fa2.js",revision:"U6_6AiTyhfp9Ki-K_78RL"},{url:"/_next/static/chunks/app/api/geocode/route-6180abf3330be1a4.js",revision:"U6_6AiTyhfp9Ki-K_78RL"},{url:"/_next/static/chunks/app/auth/callback/route-48d2cb284cec7c54.js",revision:"U6_6AiTyhfp9Ki-K_78RL"},{url:"/_next/static/chunks/app/layout-fdfd5be14737eb89.js",revision:"U6_6AiTyhfp9Ki-K_78RL"},{url:"/_next/static/chunks/app/login/page-f8b8b0790d070453.js",revision:"U6_6AiTyhfp9Ki-K_78RL"},{url:"/_next/static/chunks/app/notifications/page-66864f082d588671.js",revision:"U6_6AiTyhfp9Ki-K_78RL"},{url:"/_next/static/chunks/app/page-d0c796d547f1d5d9.js",revision:"U6_6AiTyhfp9Ki-K_78RL"},{url:"/_next/static/chunks/app/profile/%5Bid%5D/page-92ac853a5e464f30.js",revision:"U6_6AiTyhfp9Ki-K_78RL"},{url:"/_next/static/chunks/app/profile/page-29f8a979cf3639d4.js",revision:"U6_6AiTyhfp9Ki-K_78RL"},{url:"/_next/static/chunks/app/report/%5Bid%5D/page-a539c6766a076bf9.js",revision:"U6_6AiTyhfp9Ki-K_78RL"},{url:"/_next/static/chunks/app/report/new/page-e82977d58aef0ad8.js",revision:"U6_6AiTyhfp9Ki-K_78RL"},{url:"/_next/static/chunks/app/trending/page-f688d0ea98e0402e.js",revision:"U6_6AiTyhfp9Ki-K_78RL"},{url:"/_next/static/chunks/framework-6b27c2b7aa38af2d.js",revision:"U6_6AiTyhfp9Ki-K_78RL"},{url:"/_next/static/chunks/main-0ccbf0ddd675be13.js",revision:"U6_6AiTyhfp9Ki-K_78RL"},{url:"/_next/static/chunks/main-app-f83ad3163740aa35.js",revision:"U6_6AiTyhfp9Ki-K_78RL"},{url:"/_next/static/chunks/pages/_app-04c695bc05fa7935.js",revision:"U6_6AiTyhfp9Ki-K_78RL"},{url:"/_next/static/chunks/pages/_error-da21db4e7d4f6e09.js",revision:"U6_6AiTyhfp9Ki-K_78RL"},{url:"/_next/static/chunks/polyfills-42372ed130431b0a.js",revision:"846118c33b2c0e922d7b3a7676f81f6f"},{url:"/_next/static/chunks/webpack-3f1db06c145d9582.js",revision:"U6_6AiTyhfp9Ki-K_78RL"},{url:"/_next/static/css/5e8da8aef87db279.css",revision:"5e8da8aef87db279"},{url:"/_next/static/media/4cf2300e9c8272f7-s.p.woff2",revision:"18bae71b1e1b2bb25321090a3b563103"},{url:"/_next/static/media/747892c23ea88013-s.woff2",revision:"a0761690ccf4441ace5cec893b82d4ab"},{url:"/_next/static/media/8d697b304b401681-s.woff2",revision:"cc728f6c0adb04da0dfcb0fc436a8ae5"},{url:"/_next/static/media/93f479601ee12b01-s.p.woff2",revision:"da83d5f06d825c5ae65b7cca706cb312"},{url:"/_next/static/media/9610d9e46709d722-s.woff2",revision:"7b7c0ef93df188a852344fc272fc096b"},{url:"/_next/static/media/ba015fad6dcf6784-s.woff2",revision:"8ea4f719af3312a055caf09f34c89a77"},{url:"/file.svg",revision:"d09f95206c3fa0bb9bd9fefabfd0ea71"},{url:"/globe.svg",revision:"2aaafa6a49b6563925fe440891e32717"},{url:"/icons/icon.svg",revision:"4912854fbced0ec54f13367855034d23"},{url:"/jomreport.png",revision:"b08d712a0bb1ffb3aca40450f5463e54"},{url:"/manifest.json",revision:"6c8c6fa4340ae62fb166acc9c4b12b47"},{url:"/next.svg",revision:"8e061864f388b47f33a1c3780831193e"},{url:"/vercel.svg",revision:"c0af2f507b369b085b35ef4bbe3bcf1e"},{url:"/window.svg",revision:"a2760511c65806022ad20adf74370ff3"}],{ignoreURLParametersMatching:[]}),e.cleanupOutdatedCaches(),e.registerRoute("/",new e.NetworkFirst({cacheName:"start-url",plugins:[{cacheWillUpdate:async({request:e,response:i,event:s,state:a})=>i&&"opaqueredirect"===i.type?new Response(i.body,{status:200,statusText:"OK",headers:i.headers}):i}]}),"GET"),e.registerRoute(/^https:\/\/fonts\.(?:gstatic)\.com\/.*/i,new e.CacheFirst({cacheName:"google-fonts-webfonts",plugins:[new e.ExpirationPlugin({maxEntries:4,maxAgeSeconds:31536e3})]}),"GET"),e.registerRoute(/^https:\/\/fonts\.(?:googleapis)\.com\/.*/i,new e.StaleWhileRevalidate({cacheName:"google-fonts-stylesheets",plugins:[new e.ExpirationPlugin({maxEntries:4,maxAgeSeconds:604800})]}),"GET"),e.registerRoute(/\.(?:eot|otf|ttc|ttf|woff|woff2|font.css)$/i,new e.StaleWhileRevalidate({cacheName:"static-font-assets",plugins:[new e.ExpirationPlugin({maxEntries:4,maxAgeSeconds:604800})]}),"GET"),e.registerRoute(/\.(?:jpg|jpeg|gif|png|svg|ico|webp)$/i,new e.StaleWhileRevalidate({cacheName:"static-image-assets",plugins:[new e.ExpirationPlugin({maxEntries:64,maxAgeSeconds:86400})]}),"GET"),e.registerRoute(/\/_next\/image\?url=.+$/i,new e.StaleWhileRevalidate({cacheName:"next-image",plugins:[new e.ExpirationPlugin({maxEntries:64,maxAgeSeconds:86400})]}),"GET"),e.registerRoute(/\.(?:mp3|wav|ogg)$/i,new e.CacheFirst({cacheName:"static-audio-assets",plugins:[new e.RangeRequestsPlugin,new e.ExpirationPlugin({maxEntries:32,maxAgeSeconds:86400})]}),"GET"),e.registerRoute(/\.(?:mp4)$/i,new e.CacheFirst({cacheName:"static-video-assets",plugins:[new e.RangeRequestsPlugin,new e.ExpirationPlugin({maxEntries:32,maxAgeSeconds:86400})]}),"GET"),e.registerRoute(/\.(?:js)$/i,new e.StaleWhileRevalidate({cacheName:"static-js-assets",plugins:[new e.ExpirationPlugin({maxEntries:32,maxAgeSeconds:86400})]}),"GET"),e.registerRoute(/\.(?:css|less)$/i,new e.StaleWhileRevalidate({cacheName:"static-style-assets",plugins:[new e.ExpirationPlugin({maxEntries:32,maxAgeSeconds:86400})]}),"GET"),e.registerRoute(/\/_next\/data\/.+\/.+\.json$/i,new e.StaleWhileRevalidate({cacheName:"next-data",plugins:[new e.ExpirationPlugin({maxEntries:32,maxAgeSeconds:86400})]}),"GET"),e.registerRoute(/\.(?:json|xml|csv)$/i,new e.NetworkFirst({cacheName:"static-data-assets",plugins:[new e.ExpirationPlugin({maxEntries:32,maxAgeSeconds:86400})]}),"GET"),e.registerRoute(({url:e})=>{if(!(self.origin===e.origin))return!1;const i=e.pathname;return!i.startsWith("/api/auth/")&&!!i.startsWith("/api/")},new e.NetworkFirst({cacheName:"apis",networkTimeoutSeconds:10,plugins:[new e.ExpirationPlugin({maxEntries:16,maxAgeSeconds:86400})]}),"GET"),e.registerRoute(({url:e})=>{if(!(self.origin===e.origin))return!1;return!e.pathname.startsWith("/api/")},new e.NetworkFirst({cacheName:"others",networkTimeoutSeconds:10,plugins:[new e.ExpirationPlugin({maxEntries:32,maxAgeSeconds:86400})]}),"GET"),e.registerRoute(({url:e})=>!(self.origin===e.origin),new e.NetworkFirst({cacheName:"cross-origin",networkTimeoutSeconds:10,plugins:[new e.ExpirationPlugin({maxEntries:32,maxAgeSeconds:3600})]}),"GET")});
